# Add serialization writing benchmark, optimize writing