--- a/Robust.Shared/Serialization/Manager/DataDefinition/DataDefinition.Emitters.cs
+++ b/Robust.Shared/Serialization/Manager/DataDefinition/DataDefinition.Emitters.cs
@@ -96,7 +96,6 @@ namespace Robust.Shared.Serialization.Manager.DataDefinition
             return PopulateDelegate;
         }
 
-        // TODO Serialization: Turn this back into IL once it is fixed
         private SerializeDelegateSignature EmitSerializeDelegate()
         {
             MappingDataNode SerializeDelegate(
@@ -152,7 +151,6 @@ namespace Robust.Shared.Serialization.Manager.DataDefinition
             return SerializeDelegate;
         }
 
-        // TODO Serialization: Turn this back into IL once it is fixed
         // TODO Serialization: add skipHook
         private CopyDelegateSignature EmitCopyDelegate()
         {