--- a/Robust.Shared/Serialization/Markdown/Mapping/MappingDataNodeExtensions.cs
+++ b/Robust.Shared/Serialization/Markdown/Mapping/MappingDataNodeExtensions.cs
@@ -6,19 +6,19 @@ namespace Robust.Shared.Serialization.Markdown.Mapping
 {
     public static class MappingDataNodeExtensions
     {
-        public static MappingDataNode AddNode(this MappingDataNode mapping, string key, DataNode node)
+        public static MappingDataNode Add(this MappingDataNode mapping, string key, DataNode node)
         {
             mapping.Add(new ValueDataNode(key), node);
             return mapping;
         }
 
-        public static MappingDataNode AddNode(this MappingDataNode mapping, string key, string value)
+        public static MappingDataNode Add(this MappingDataNode mapping, string key, string value)
         {
             mapping.Add(new ValueDataNode(key), new ValueDataNode(value));
             return mapping;
         }
 
-        public static MappingDataNode AddNode(this MappingDataNode mapping, string key, List<string> sequence)
+        public static MappingDataNode Add(this MappingDataNode mapping, string key, List<string> sequence)
         {
             mapping.Add(new ValueDataNode(key), new SequenceDataNode(sequence));
             return mapping;