--- a/Robust.UnitTesting/Shared/Serialization/TypeSerializers/DictionarySerializerTest.cs
+++ b/Robust.UnitTesting/Shared/Serialization/TypeSerializers/DictionarySerializerTest.cs
@@ -40,9 +40,9 @@ namespace Robust.UnitTesting.Shared.Serialization.TypeSerializers
             };
             var node = new MappingDataNode();
 
-            node.AddNode("1", new ValueDataNode("A"));
-            node.AddNode("2", new ValueDataNode("B"));
-            node.AddNode("3", new ValueDataNode("C"));
+            node.Add("1", new ValueDataNode("A"));
+            node.Add("2", new ValueDataNode("B"));
+            node.Add("3", new ValueDataNode("C"));
 
             var deserializedDictionary = Serialization.ReadValue<Dictionary<int, string>>(node);