Before
|                          Method |      Mean |     Error |    StdDev |
|-------------------------------- |----------:|----------:|----------:|
|                     WriteString |  3.234 us | 0.0643 us | 0.1399 us |
|                    WriteInteger |  4.524 us | 0.0848 us | 0.1770 us |
|   WriteDataDefinitionWithString |  7.985 us | 0.1439 us | 0.1275 us |
|         WriteSeedDataDefinition | 91.377 us | 1.5887 us | 1.4861 us |
| BaselineWriteSeedDataDefinition |  1.959 us | 0.0231 us | 0.0193 us |


After
|                          Method |       Mean |    Error |   StdDev |     Median |
|-------------------------------- |-----------:|---------:|---------:|-----------:|
|                     WriteString |   136.3 ns |  2.68 ns |  7.05 ns |   133.8 ns |
|                    WriteInteger |   145.2 ns |  2.97 ns |  6.76 ns |   143.6 ns |
|   WriteDataDefinitionWithString |   568.7 ns | 11.32 ns | 16.23 ns |   562.5 ns |
|         WriteSeedDataDefinition | 9,274.0 ns | 85.83 ns | 67.01 ns | 9,276.8 ns |
| BaselineWriteSeedDataDefinition | 1,923.2 ns | 35.99 ns | 31.91 ns | 1,912.7 ns |